+++ packages/execution-grid-client/test/it/proxy.spec.ts
@@ -2,6 +2,7 @@ import assert from 'assert'
 import {createServer as createHttpServer, Server as HttpServer} from 'http'
 import fetch from 'node-fetch'
 import {makeLogger} from '@applitools/logger'
+import {modifyIncomingMessage, type ModifiedIncomingMessage} from '../../src/incoming-message'
 import {makeProxy} from '../../src/proxy'
 
 describe('proxy', () => {
@@ -35,7 +36,8 @@ describe('proxy', () => {
           }
         })
 
-        proxyServer.on('request', async (request, response) => {
+        proxyServer.on('request', async (message: ModifiedIncomingMessage, response) => {
+          const request = modifyIncomingMessage(message)
           try {
             headers.original = request.headers
             await proxyRequest({request, response, options: {url: 'http://localhost:3000'}, logger})
@@ -81,7 +83,8 @@ describe('proxy', () => {
           }
         })
 
-        secondProxyServer.on('request', async (request, response) => {
+        secondProxyServer.on('request', async (message: ModifiedIncomingMessage, response) => {
+          const request = modifyIncomingMessage(message)
           try {
             headers.proxied = request.headers
             await proxyRequest({request, response, logger})
@@ -90,7 +93,8 @@ describe('proxy', () => {
           }
         })
 
-        proxyServer.on('request', async (request, response) => {
+        proxyServer.on('request', async (message, response) => {
+          const request = modifyIncomingMessage(message)
           try {
             headers.original = request.headers
             await proxyRequest({
@@ -118,6 +122,82 @@ describe('proxy', () => {
         server.close()
         proxyServer.close()
         secondProxyServer.close()
+      }
+    })
+  })
+
+  it('retries with body', async () => {
+    return new Promise<void>(async (resolve, reject) => {
+      const proxyRequest = makeProxy({shouldRetry: response => response.statusCode >= 400, retryTimeout: 0})
+      const server = await createServer({port: 3000})
+      const proxyServer = await createServer({port: 4000})
+      try {
+        let requestCount = 0
+        server.on('request', async (message, response) => {
+          const request = modifyIncomingMessage(message)
+          if (requestCount < 5) {
+            requestCount += 1
+            response.writeHead(500).end()
+          } else {
+            response.writeHead(200, {'Content-Type': request.headers['content-type']}).end(await request.body())
+          }
+        })
+
+        proxyServer.on('request', async (message, response) => {
+          const request = modifyIncomingMessage(message)
+          try {
+            await proxyRequest({request, response, options: {url: 'http://localhost:3000'}, logger})
+          } catch (err) {
+            reject(err)
+          }
+        })
+
+        const body = {value: true}
+
+        const response = await fetch('http://localhost:4000/path', {method: 'post', body: JSON.stringify(body)})
+
+        assert.strictEqual(response.status, 200)
+        assert.deepStrictEqual(await response.json(), body)
+        resolve()
+      } catch (err) {
+        reject(err)
+      } finally {
+        server.close()
+        proxyServer.close()
+      }
+    })
+  })
+
+  it('respond on failed requests with body', async () => {
+    return new Promise<void>(async (resolve, reject) => {
+      const proxyRequest = makeProxy({
+        shouldRetry: async response => response.statusCode >= 400 && (await response.json())?.error,
+        retryTimeout: 0,
+      })
+      const server = await createServer({port: 3000})
+      const proxyServer = await createServer({port: 4000})
+      try {
+        server.on('request', async (_, response) => response.writeHead(404).end(JSON.stringify({error: 'blabla'})))
+
+        proxyServer.on('request', async (message, response) => {
+          const request = modifyIncomingMessage(message)
+          try {
+            await proxyRequest({request, response, options: {url: 'http://localhost:3000'}, logger})
+          } catch (err) {
+            reject(err)
+          }
+        })
+
+        const response = await fetch('http://localhost:4000/path')
+
+        assert.strictEqual(response.status, 404)
+        assert.deepStrictEqual(await response.json(), {error: 'blabla'})
+        resolve()
+      } catch (err) {
+        reject(err)
+      } finally {
+        server.close()
+        proxyServer.close()
       }
     })
   })