+++ packages/eyes-sdk-core/test/unit/server/ServerConnector.spec.js
@@ -1,8 +1,7 @@
-'use strict'
-
 const assert = require('assert')
 const assertRejects = require('assert-rejects')
 const settle = require('axios/lib/core/settle')
+const fs = require('fs')
 const {startFakeEyesServer} = require('@applitools/sdk-fake-eyes-server')
 const {makeLogger} = require('@applitools/logger')
 const {
@@ -829,5 +828,30 @@ render height & width are required when deviceEmulationInfo is not provided, req
       ),
     )
     assert.ok(renderErr.message.includes(`Error: combination of url, dom, resources is invalid`))
+  })
+
+  it('works with self signed certificates', async () => {
+    const {port, close} = await startFakeEyesServer({
+      cert: fs.readFileSync('./test/fixtures/certificate.pem'),
+      key: fs.readFileSync('./test/fixtures/key.pem'),
+      logger,
+    })
+    const serverUrl = `https://localhost:${port}`
+    const serverConnector = getServerConnector({serverUrl})
+
+    try {
+      await serverConnector.startSession(
+        new SessionStartInfo({
+          appIdOrName: 'app id or name',
+          scenarioIdOrName: 'scenario id or name',
+          agentId: 'agent id',
+          batchInfo: {name: 'batch name'},
+          environment: {os: 'os', hostingApp: 'hosting app', displaySize: {width: 1.5, height: 1.5}},
+          defaultMatchSettings: {},
+        }),
+      )
+    } finally {
+      await close()
+    }
   })
 })